requirements.txt
absl-py==2.3.1
altair==5.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==3.4.3
ciecplib==0.8.0
click==8.1.7
cmake==3.29.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.7
curl_cffi==0.13.0
cycler==0.12.1
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-bootstrap-templates==1.1.2
dash-core-components==2.0.0
dash-cytoscape==1.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dcc==0.8.0
decorator==5.1.1
Django==5.0.4
download==0.3.5
executing==2.0.1
flak==0.1.5
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==25.2.10
fonttools==4.50.0
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.74.0
h11==0.14.0
h5py==3.14.0
html2text==2024.2.26
humanfriendly==10.0
idna==3.10
importlib_metadata==7.1.0
iniconfig==2.0.0
interchange==2021.0.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonpickle==3.0.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.11.3
kiwisolver==1.4.5
libclang==18.1.1
logger==1.4
lxml==5.2.2
Markdown==3.9
markdown-it-py==4.0.0
markup==0.2
MarkupSafe==3.0.2
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
monotonic==1.6
multitasking==0.0.12
mypy-extensions==1.0.0
namex==0.1.0
neo4j==5.18.0
nest-asyncio==1.6.0
network==0.1
networkx==3.2.1
numpy==1.26.0
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.17.0
outcome==1.2.0
output==1.0.1
packaging==23.2
pandas==2.2.1
pansi==2020.7.3
parso==0.8.4
pathspec==0.12.1
peewee==3.18.2
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.8
pure-eval==0.2.2
py2neo==2021.2.4
pyarrow==15.0.2
pycparser==2.21
pydeck==0.8.1b0
pydot==2.0.0
pydotplus==2.0.2
Pygments==2.19.2
pyngrok==7.1.6
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.1
referencing==0.34.0
render==1.0.0
requests==2.32.5
requests-ecp==0.3.1
retrying==1.3.4
rich==13.9.4
rpds-py==0.18.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==4.6.0
sequential==1.0.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
state==0.1.1.dev0
streamlit==1.32.2
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==3.1.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.5.0
watchdog==4.0.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
xgboost==3.0.4
yfinance==0.2.65
zipp==3.18.1
